# cmd/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for Gym Management Platform",
        "title": "BRO Gateway API",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticate staff member and get JWT token. If user belongs to multiple companies, returns a company selector.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Staff login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/staff.LoginResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CompanyStaffOption": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "company_slug": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.LoginMultiCompanyResponse": {
            "type": "object",
            "properties": {
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.CompanyStaffOption"
                    }
                },
                "requires_company_selection": {
                    "type": "boolean"
                },
                "temp_token": {
                    "description": "Short-lived token for company selection",
                    "type": "string"
                },
                "temp_token_expires_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "domain.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 1
                }
            }
        },
        "domain.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "staff": {
                    "$ref": "#/definitions/domain.StaffResponse"
                }
            }
        },
...
        "domain.StaffResponse": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "company_slug": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "errs.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "staff.LoginResult": {
            "type": "object",
            "properties": {
                "direct_login": {
                    "description": "If single company, these are set",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.LoginResponse"
                        }
                    ]
                },
                "multi_company": {
                    "description": "If multiple companies, these are set",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.LoginMultiCompanyResponse"
                        }
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}